requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.24.2
pandas==2.0.0
openpyxl==3.1.2
sentence-transformers==2.2.2
huggingface-hub==0.10.1
nltk==3.7
scikit-learn==1.0.2
torch==2.0.0
websockets>=10.0  # Ajout pour le support WebSocket d'Uvicorn
aiofiles

# Nouvelles dépendances pour la visualisation et le PageRank
networkx>=2.8.0  # Pour le calcul du PageRank et la manipulation de graphes
matplotlib>=3.5.0  # Pour générer des visualisations statiques si nécessaire

# Dépendances pour le web crawling
scrapy>=2.8.0  # Framework de crawling robuste
beautifulsoup4>=4.12.0  # Pour l'extraction de contenu
requests>=2.28.0  # Pour les requêtes HTTP simples
asyncio>=3.4.3  # Pour la gestion asynchrone

# Dépendances pour l'intégration Google Search Console
google-auth>=2.16.0  # Authentification Google
google-auth-oauthlib>=1.0.0  # Support OAuth pour l'authentification
google-auth-httplib2>=0.1.0  # Adapter HTTP pour l'authentification
google-api-python-client>=2.70.0  # Client API pour Google Search Console

# Dépendances pour la gestion des sessions
fastapi-sessions>=0.3.2  # Gestion des sessions pour FastAPI
itsdangerous>=2.1.2  # Cryptographie pour les sessions
python-multipart>=0.0.5  # Pour le parsing des form-data
tqdm>=4.65.0  # Pour les barres de progression